Create subset now supports dataset to be created hidden Progress section and percentage is working with data hierarchy item

+++ HDPS/src/DataHierarchyItem.h
@@ -114,7 +114,7 @@ signals:
      * Signals that the progress changed
      * @param progress Progress
      */
-    void progressChanged(const double& progress);
+    void progressChanged(const float& progress);
 
 protected:
